+++ src/Mock/AbstractGuiStubBuilder.ts
@@ -5,6 +5,13 @@ export abstract class AbstractGuiStubBuilder implements BuilderInterface {
     protected source: string;
     protected request = require("sync-request");
     protected jsdom = require("jsdom");
+    protected location: string;
+
+    public constructor() {
+        let Window = require("jsdom/lib/jsdom/browser/Window");
+        global["Window"] = Window;
+        require("jsdom-global")();
+    }
 
     public abstract getMock();
 
@@ -15,5 +22,9 @@ export abstract class AbstractGuiStubBuilder implements BuilderInterface {
 
     protected createDom(markup: string, config: Object = {}): Document {
         return this.jsdom.jsdom(markup, config);
+    }
+
+    public createNewWindow(): Window {
+        return new global["Window"]({ parsingMode: "html" , "url": this.location});
     }
 }