+++ tests/Mock/WindowBuilderTest.ts
@@ -6,6 +6,7 @@ export class WindowBuilderTest extends AbstractUnitTestCase {
     public testStubWindowFromUrl() {
         let builder: WindowBuilder = new WindowBuilder();
         let window: Window = builder.setLocation("http://example.com", true).getMock();
+        this.assertInstanceOf(Window, window);
         this.assertEquals("example.com", window.location.host);
         this.assertEquals("Example Domain", window.document.title);
     }